some final cleanup
