--- grpc/grpc-specs/workflows/templates/README.md.template
@@ -1,15 +1,0 @@
-# {{PACKAGE_NAME}}
-
-This package contains Protocol Buffer definitions and generated TypeScript code for RPC services.
-
-## Prerequisites
-
-You need the Protocol Buffers compiler (protoc) installed to generate TypeScript code from the proto definitions:
-
-- macOS: `brew install protobuf`
-- Linux: `apt-get install protobuf-compiler`
-- Windows: Download from [protobuf releases](https://github.com/protocolbuffers/protobuf/releases)
-
-## Usage
-
-Run `npm run generate` to generate TypeScript code from the proto definitions.
