--- grpc/grpc-specs/workflows/templates/example.proto.template
@@ -1,20 +1,0 @@
-syntax = "proto3";
-
-package {{PROTO_PACKAGE_NAME}};
-
-import "envelope.proto";
-
-service ExampleService {
-  rpc ExampleMethod(ExampleRequest) returns (ExampleResponse);
-}
-
-message ExampleRequest {
-  saflib.common.v1.SafAuth auth = 1;
-  saflib.common.v1.SafRequest request = 2;
-  string example_field = 3;
-}
-
-message ExampleResponse {
-  bool successful = 1;
-  string message = 2;
-}
