--- grpc/grpc-specs/workflows/templates/index.ts.template
@@ -1,2 +1,0 @@
-// Generated exports will be added here after running npm run generate
-// Example: export * from "./dist/example.ts";
