--- grpc/grpc-specs/workflows/templates/tsconfig.json.template
@@ -1,3 +1,0 @@
-{
-  "extends": "@saflib/monorepo/tsconfig.json"
-}
