+++ workflows/src/workflow.ts
@@ -1,12 +1,5 @@
-import type {
-  CLIArgument,
-  ChecklistItem,
-  Step,
-  WorkflowBlob,
-  WorkflowStatus,
-} from "./types.ts";
+import type { CLIArgument, ChecklistItem, WorkflowBlob } from "./types.ts";
 import type { AnyStateMachine, AnyActor } from "xstate";
-import { addNewLinesToString } from "@saflib/utils";
 import { createActor, waitFor } from "xstate";
 import { getSafReporters } from "@saflib/node";
 import path from "node:path";
@@ -18,7 +11,8 @@ import type {
   WorkflowOutput,
 } from "./xstate.ts";
 import type { ReturnsError } from "@saflib/monorepo";
-// The following is TS magic to describe a class constructor that implements the abstract SimpleWorkflow class.
+
+// The following is TS magic to describe a class constructor that implements the Workflow class.
 type AbstractClassConstructor<T extends Workflow> = new (...args: any[]) => T;
 
 /**
@@ -39,8 +33,7 @@ export interface WorkflowMeta {
 }
 
 /**
- * Abstract superclass for SimpleWorkflow and XStateWorkflow. To be removed
- * when XStateWorkflow is fully adopted.
+ * Abstract superclass for XStateWorkflow. Can probably be removed since SimpleWorkflows are gone.
  */
 export abstract class Workflow {
   abstract readonly name: string;
@@ -57,145 +50,6 @@ export abstract class Workflow {
   abstract done(): boolean;
   abstract getChecklist(): ChecklistItem[];
   abstract getError(): Error | undefined;
-}
-
-/**
- * First iteration of workflows. Opted to try using XState instead for plenty
- * of built in FSM features and tooling.
- *
- * @deprecated Use XStateWorkflow instead.
- */
-export abstract class SimpleWorkflow<
-  P extends Record<string, any>,
-  D extends Record<string, any> = {},
-> extends Workflow {
-  params?: P;
-  data?: D;
-
-  abstract init: (...args: any[]) => Promise<ReturnsError<D>>;
-  abstract steps: Step[];
-  abstract workflowPrompt: () => string;
-  private stepIndex = 0;
-  private status: WorkflowStatus = "not started";
-
-  getData(): D {
-    if (!this.data) {
-      throw new Error(`Data is not set for workflow ${this.constructor.name}`);
-    }
-    return this.data;
-  }
-
-  getParams(): P {
-    if (!this.params) {
-      throw new Error(
-        `Params are not set for workflow ${this.constructor.name}`,
-      );
-    }
-    return this.params;
-  }
-
-  setData(data: D) {
-    this.data = data;
-  }
-
-  print(message: string) {
-    const { log } = getSafReporters();
-    log.info("");
-    log.info(addNewLinesToString(message));
-  }
-
-  async kickoff(): Promise<boolean> {
-    if (this.status === "completed") {
-      return true;
-    }
-    this.status = "in progress";
-    this.print(`The "${this.name}" workflow has been kicked off.`);
-    await this.printStatus();
-    this.print(`To continue, run "npm exec saf-workflow next"`);
-    return true;
-  }
-
-  async printStatus() {
-    if (this.status === "completed") {
-      this.print("The workflow has been completed.");
-      return;
-    }
-    if (this.status === "not started") {
-      this.print("The workflow has not started yet.");
-      return;
-    }
-    this.print(this.workflowPrompt());
-    this.print(this.steps[this.stepIndex].prompt());
-  }
-
-  async goToNextStep() {
-    if (this.status === "completed") {
-      this.print("The workflow has already been completed.");
-      return;
-    }
-    this.stepIndex++;
-    if (this.stepIndex >= this.steps.length) {
-      this.status = "completed";
-      this.print(`The workflow has been completed.`);
-      return;
-    }
-    this.print(
-      `The workflow has moved to step "${this.steps[this.stepIndex].name}".`,
-    );
-    await this.printStatus();
-  }
-
-  dehydrate(): WorkflowBlob {
-    return {
-      workflowName: this.name,
-      internalState: {
-        status: this.status,
-        stepIndex: this.stepIndex,
-        data: this.data ?? {},
-        params: this.params ?? {},
-      },
-    };
-  }
-
-  hydrate(blob: WorkflowBlob): void {
-    if (typeof blob !== "object" || blob === null) {
-      throw new Error("Invalid serialized data: not an object");
-    }
-    if (!blob.internalState) {
-      throw new Error("Invalid serialized data: no internal state");
-    }
-    if (
-      !["not started", "in progress", "completed"].includes(
-        blob.internalState.status,
-      )
-    ) {
-      throw new Error(`Invalid status: ${blob.internalState.status}`);
-    }
-    if (typeof blob.internalState.stepIndex !== "number") {
-      throw new Error("Invalid stepIndex: not a number");
-    }
-    this.params = blob.internalState.params as any;
-    this.data = blob.internalState.data as any;
-    this.stepIndex = blob.internalState.stepIndex;
-    this.status = blob.internalState.status;
-  }
-
-  done = (): boolean => {
-    return this.status === "completed";
-  };
-
-  getChecklist = (): ChecklistItem[] => {
-    return [];
-  };
-
-  getCurrentStateName = (): string => {
-    const currentStep = this.steps[this.stepIndex];
-    return currentStep.name;
-  };
-
-  getError = (): Error | undefined => {
-    return undefined;
-  };
 }
 
 interface XStateWorkflowOptions {
